Drop data and throw exception if sending fails
Previously:
The reconnect logic in getOutConnOrOpen would only engage in case
connecting to the given address failed. If the connection is made, but
is later severed, then the socket frame's output queue will hang around
and retain the data that previously failed to send.

Today I observed (and continue to observe as I write this) the cardano
smart generator log warnings of massive (multiple 1000s of seconds)
delays in sending messages (such as Send Txs to neighbors). This is
coupled with ludicrous heap size (6GB and growing). My theory is that
all of the data which was to be sent is lingering in the egress queues
for various socket frames which can't be cleared.

This patch may fix it.