sfSend unblocked by job curator
Previously, if the job curator is not finished when sfSend is called,
it will block until the notifier is called (which it places into its
egress channel). Thus if the notifier is never called, it blocks
indefinitely. The notifier will presumably be called by sfProcessSocket,
but not if there's an exception on the sending sink (for instance if the
connection is severed). When such a thing happens, the SocketFrame is
released and the job curator is closed. With this patch, a closing of
the job curator will also unblock sfSend. The user of sfSend will not be
made aware of whether or not their data went through.