TCP_USER_TIMEOUT 3s, stop retry after 3 exceptions
It is observed that sometimes a connection will remain open
indefinitely, and the egress queues will fill up. I am still trying to
determine exactly why this happens, but my theory is that the peer makes
a connection to the local host while the local host's connection to that
same peer is temporarily down (waiting 3 seconds to reconnect). After
the reconnection, could it be that the peer never receives from that
socket, because it has another one with the local host? TBD

Anyway, I will run the experiment with this revision, in which socket
writes will timeout after 3 seconds. This should eliminate the egress
queue choking, for after 3 seconds a reconnect attempt will be made, and
if the problem continues, the socket will be dropped after 3
reconnections, for an estimated total time of 6 * 3 = 18 seconds.